refactoring calculator service to return a random equation & answer
